==> 04b_stocks_recover_clean.Rmd <==
---
title: "Identifying Price Informativeness"
author: "Eduardo Davila^[Yale] & Cecilia Parlatore^[NYU Stern]"
date: "`r Sys.Date()`"
output:
  html_document: default
  pdf_document: default
---

```{r echo=FALSE, include=FALSE}

library(here); library(tidyverse); library(foreach); library(doParallel); library(AER)
path <- here::here(); print(path); setwd(path); rm(path)

source("functions/fn_remove_outliers.R")

load("intermediate/data_selected.RData")
load("intermediate/results_rolling.RData")
load("intermediate/results_full.RData")

out_threshold     <- 0.95
outlier           <- TRUE
contiguous        <- TRUE

```

We can filter stocks based on the highest *leverage score*.
We can filter stocks so that they have contiguous observations

```{r}

if (outlier == TRUE) {
  
  results_rolling_q_outlier <- remove_outliers(results_rolling_q, out_threshold)
  results_rolling_a_outlier <- remove_outliers(results_rolling_a, out_threshold)
  results_full_q_outlier    <- remove_outliers(results_full_q,    out_threshold)
  results_full_a_outlier    <- remove_outliers(results_full_a,    out_threshold)
  
}

if (contiguous == TRUE) {
  
  results_rolling_q_clean <- results_rolling_q_outlier %>%
    mutate(datefrac_start = year_start + month_start/12,
           datefrac_end   = year_end   + month_end/12, 
           date_dif       = datefrac_end - datefrac_start) %>% 
    filter(date_dif == N_q/4 - 0.25)
  
  results_rolling_a_clean <- results_rolling_a_outlier %>%
    mutate(datefrac_start = year_start + month_start/12,
           datefrac_end   = year_end   + month_end/12, 
           date_dif       = datefrac_end - datefrac_start) %>% 
    filter(date_dif == N_a - 1)
}

n_permno_q_before_outlier_q <- length(unique(results_rolling_q$permno))
n_permno_q_after_outlier_q  <- length(unique(results_rolling_q_outlier$permno))
n_permno_q_clean            <- length(unique(results_rolling_q_clean$permno))

n_permno_a_before_outlier_a <- length(unique(results_rolling_a$permno))
n_permno_a_after_outlier_a  <- length(unique(results_rolling_a_outlier$permno))
n_permno_a_clean            <- length(unique(results_rolling_a_clean$permno))

results_rolling_q <- results_rolling_q_clean
results_rolling_a <- results_rolling_a_clean

save(results_rolling_q, results_rolling_a, file = "intermediate/results_rolling_clean.RData")

```

```{r}

if (contiguous == TRUE) {
  
  jump_q <- df_q %>%
    group_by(permno) %>%
    mutate(date_frac = year + month/12) %>%
    mutate(date_dif = date_frac - lag(date_frac)) %>%
    summarize(jump_max   = max(date_dif,        na.rm = T),
              jump_min   = min(date_dif,        na.rm = T),
              jump_count = sum(date_dif > 0.25, na.rm = T), .groups = "drop_last")
  
  jump_q_clean <- jump_q %>% filter(jump_max == 0.25, jump_min == 0.25, jump_count == 0)

  jump_a <- df_a %>%
    group_by(permno) %>%
    mutate(year_dif = year - lag(year)) %>% 
    summarize(jump_max   = max(year_dif,     na.rm = T), 
              jump_min   = min(year_dif,     na.rm = T),
              jump_count = sum(year_dif > 1, na.rm = T), .groups = "drop_last") 
  
  jump_a_clean <- jump_a %>% filter(jump_max == 1, jump_min == 1, jump_count == 0)

  results_full_q_clean <- results_full_q_outlier %>% filter(permno %in% jump_q_clean$permno)
  results_full_a_clean <- results_full_a_outlier %>% filter(permno %in% jump_a_clean$permno)
  
}

results_full_q <- results_full_q_clean
results_full_a <- results_full_q_clean

save(results_full_q, results_full_a, jump_q, jump_a, file = "intermediate/results_full_clean.RData")

```
